+++ CHANGELOG.md
@@ -9,3 +9,4 @@ master
   * throw \RuntimeException to make test fail
   * make sure element is visible or not
 * Enable CS rule to add backslash for compiler-optimized functions
+* Id ending popin selector is now required for Sonata popin rules